Provide POST parameters to actions
Read and parse POST method parameters and add to the request.params object
available to actions. Made it so it would parse parameters of the form
param[subparam] into an appropriate javascript object. Added example code
following a rest framework for a micropost new/create page pair.

Fixed problem in view spec where not spying on the asyncrhonous readfile
later caused problems with the asynchronus test for the POST parameter
handling (this is the first use of async jasmine testing).