Nails script additions; Packaged for npm
Moved nails program to bin and changed it to accept either a new or
server command. New now creates a new nails application skeleton. Server
starts running the nails server when in the home directory of an
application. Also changed nails to be a coffee script program so now
a separate install of coffee-script appears to be necessary (this is
because npm doesn't install modules with bin correctly as a dependancy).

Moved the code that was being used to demonstrate various aspects of the
server to the example directory. Because it is now a directory lower,
had to change the about page to point to the README in the home
directory.

By cleaning up the directories and adding nails new we are now able to
release this as an npm module. Added package.json to describe the
nailsjs module.

Added the proto directory to contain a prototype of the directory
structure that nails new creates.

Changed the name of the nails directory to be lib.

Updated the README to contain instructions for using the npm installer
and how to create new apps and run them with the nails server command.

Removed a dependency on jade from the view library.